Use fixed size types

+++ math.go
@@ -3,7 +3,7 @@ package main
 import "C"
 
 //export add
-func add(x int, y int) int {
+func add(x int64, y int64) int64 {
 	return x + y
 }
